Move to Therapies, working on Swagger (draft)
